requirements.txt
APScheduler==3.6.3
Django==2.2.2
matplotlib==3.2.1
numpy==1.18.2
paho-mqtt==1.5.0
Pillow==7.1.1
sklearn==0.0
